Refactor interview groups to dynamic route and fix split panel widths
