+++ components/icons/check-circle.icon.tsx
@@ -1,0 +1,23 @@
+import { ReactElement } from "react";
+
+const CheckCircleIcon = (): ReactElement => {
+  return (
+    <svg
+      width="22"
+      height="17"
+      viewBox="0 0 22 17"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M20 2L8 14L2 8"
+        stroke="currentColor"
+        strokeWidth="2.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+};
+
+export default CheckCircleIcon;
